move typing.hpp to common
