+++ tests/cpp_unit_tests/test_typing.cpp
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: MPL-2.0
 
 #include <power_grid_model/common/power_grid_model.hpp>
-#include <power_grid_model/typing.hpp>
+#include <power_grid_model/common/typing.hpp>
 
 #include <doctest/doctest.h>
